[4] moving the packages and preparing for new service
